EZ-2934: EZ-2937 Default accounts: apply the same logic as in the auto-rejection rules for defaults - take 3

+++ Tools/EzbobUtils/AutomationVerification/AutomationVerification/AutoDecision/AutoApproval/Configuration.cs
@@ -36,7 +36,7 @@
 		public virtual int HmrcTurnoverAge { get; set; }
 		public virtual decimal HmrcTurnoverDropQuarterRatio { get; set; }
 		public virtual decimal HmrcTurnoverDropHalfYearRatio { get; set; }
-		public virtual int Reject_Defaults_Amount { get; set; }
+		public virtual decimal Reject_Defaults_Amount { get; set; }
 		public virtual int Reject_Defaults_MonthsNum { get; set; }
 
 		public virtual string AllowedCaisStatusesWithLoan {
